agroai/agro-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scikit-learn>=1.4.2
joblib==1.4.0
numpy>=1.26.4
pandas==2.2.2
pydantic==2.7.1
requests==2.31.0
python-dotenv==1.0.1
httpx==0.27.0
